Enforce presence of options for some commands :policewoman:
